+++ metrics/tarantool/network.lua
@@ -10,11 +10,11 @@ local function update_network_metrics()
     local box_stat_net = box.stat.net()
 
     collectors_list.net_sent_total =
-        utils.set_gauge('net_sent_total', 'Totally sent in bytes', box_stat_net.SENT.total)
+        utils.set_counter('net_sent_total', 'Totally sent in bytes', box_stat_net.SENT.total)
     collectors_list.net_sent_rps =
         utils.set_gauge('net_sent_rps', 'Sending RPS', box_stat_net.SENT.rps)
     collectors_list.net_received_total =
-        utils.set_gauge('net_received_total', 'Totally received in bytes', box_stat_net.RECEIVED.total)
+        utils.set_counter('net_received_total', 'Totally received in bytes', box_stat_net.RECEIVED.total)
     collectors_list.net_received_rps =
         utils.set_gauge('net_received_rps', 'Receive RPS', box_stat_net.RECEIVED.rps)
 
@@ -25,7 +25,7 @@ local function update_network_metrics()
         collectors_list.net_connections_rps =
             utils.set_gauge('net_connections_rps', 'Connection RPS', box_stat_net.CONNECTIONS.rps)
         collectors_list.net_connections_total =
-            utils.set_gauge('net_connections_total', 'Connections total amount', box_stat_net.CONNECTIONS.total)
+            utils.set_counter('net_connections_total', 'Connections total amount', box_stat_net.CONNECTIONS.total)
         collectors_list.net_connections_current =
             utils.set_gauge('net_connections_current', 'Current connections amount', box_stat_net.CONNECTIONS.current)
     elseif box_stat_net.CONNECTIONS ~= nil then
@@ -37,7 +37,7 @@ local function update_network_metrics()
         collectors_list.net_requests_rps =
             utils.set_gauge('net_requests_rps', 'Requests RPS', box_stat_net.REQUESTS.rps)
         collectors_list.net_requests_total =
-            utils.set_gauge('net_requests_total', 'Requests total amount', box_stat_net.REQUESTS.total)
+            utils.set_counter('net_requests_total', 'Requests total amount', box_stat_net.REQUESTS.total)
         collectors_list.net_requests_current =
             utils.set_gauge('net_requests_current', 'Pending requests', box_stat_net.REQUESTS.current)
     end