+++ metrics/tarantool/vinyl.lua
@@ -59,6 +59,8 @@ local function update()
         vinyl_stat.scheduler.dump_time)
     collectors_list.vinyl_scheduler_dump_count =
         utils.set_gauge('vinyl_scheduler_dump_count', 'The count of completed dumps', vinyl_stat.scheduler.dump_count)
+    collectors_list.vinyl_scheduler_dump_total =
+        utils.set_counter('vinyl_scheduler_dump_total', 'The count of completed dumps', vinyl_stat.scheduler.dump_count)
 end
 
 return {