+++ CHANGELOG.md
@@ -7,7 +7,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
+
 - Don't reset collectors when Cartridge roles hot reload
+
+### Changed
+
+- Type changed from `gauge` to `counter`:
+  - `tnt_net_sent_total`
+  - `tnt_net_received_total`
+  - `tnt_net_connections_total`
+  - `tnt_net_requests_total`
+  - `tnt_stats_op_total`
+
+### Added
+
+- New metrics:
+  - `tnt_vinyl_tuples` (same as `tnt_space_count`)
+  - `tnt_fiber_amount` (same as `tnt_fiber_count`)
+  - `lj_gc_memory` (same as `lj_gc_total`)
+  - `tnt_cpu_number` (same as `tnt_cpu_count`)
+  - `tnt_cpu_time` (same as `tnt_cpu_total`)
+  - `tnt_vinyl_scheduler_dump_total` (same as `tnt_vinyl_scheduler_dump_count`)
 
 ### Deprecated
 
@@ -17,6 +37,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `tnt_net_connections_rps`
   - `tnt_net_requests_rps`
   - `tnt_stats_op_rps`
+  - `tnt_space_count`
+  - `tnt_fiber_count`
+  - `lj_gc_total`
+  - `tnt_cpu_count`
+  - `tnt_cpu_total`
+  - `tnt_vinyl_scheduler_dump_count`
 
 ## [0.12.0] - 2021-11-18
 ### Changed